requirements.txt
# AEGIS Backend Dependencies
classified-agent>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
python-dotenv>=1.0.0
pytest>=7.0.0
pydantic>=2.0.0
web3>=6.0.0
